re-factory to use gameFactory
