+++ app/app.js
@@ -4,8 +4,7 @@
 	var canvasService = require('services/canvas.service');
 	var documentService = require('services/document.service');
 	var eventService = require('services/event.service');
-	var game = require('components/game');
-	var grid = require('components/grid');
+	var gameFactory = require('services/game.factory');
 
 	module.exports = {
 		init: function() {
@@ -14,25 +13,18 @@
 			function onDOMLoaded() {
 				console.log('DOM has loaded.');
 				var context = canvasService.contextFactory('grid');
-				grid.drawGrid(context());
+				var game = gameFactory.game(context);
+				game.newGame();
+
 				eventService.addEventListener(documentService.getElementById('grid'), 'click', gridClick);
 				eventService.addEventListener(documentService.getElementById('newGame'), 'click', newGame);
 
 				function gridClick(event) {
-					var cell = grid.onClick(event);
-					if (cell !== undefined) {
-						if (game.isValidMove(cell)) {
-							console.log('row: ' + cell.row.id + ' column: ' + cell.column.id);
-							game.move(cell);
-							grid.drawO(context(), cell);
-							grid.drawX(context(), cell);
-						}
-					}
+					game.move(event);
 				}
 
 				function newGame(event) {
 					game.newGame();
-					grid.newGame(context());
 				}
 			}
 		}
